made the tutorial it's own component
